Allow using UBSan's unsigned-overflow check
In C and C++ unsigned overflow has defined behaviour (unlike signed
overflow) but in many cases it happening is not intended and indicates
a bug.

UBSan supports an unsigned-overflow check (at least the clang version
does - currently the GCC version doesn't seem to) so I've annotated the
few places in the code where we deliberately rely on unsigned overflow
(and eliminated a few cases where we can easily not rely on it) and
added this check to the clang-sanitisers job.

+++ xapian-core/include/xapian/unicode.h
@@ -392,7 +392,7 @@ inline unsigned tolower(unsigned ch) {
     int info = Xapian::Unicode::Internal::get_character_info(ch);
     if (!(info & Internal::INFO_TOLOWER_MASK))
 	return ch;
-    return ch + Internal::get_delta(info);
+    return unsigned(int(ch) + Internal::get_delta(info));
 }
 
 /// Convert a Unicode character to uppercase.
@@ -400,7 +400,7 @@ inline unsigned toupper(unsigned ch) {
     int info = Xapian::Unicode::Internal::get_character_info(ch);
     if (!(info & Internal::INFO_TOUPPER_MASK))
 	return ch;
-    return ch - Internal::get_delta(info);
+    return unsigned(int(ch) - Internal::get_delta(info));
 }
 
 /// Convert a UTF-8 std::string to lowercase.
